fix collision add Gravity add comment for som bug along the way re-modify the KeyboardInputs.java to read multiple input buttons as one some bug happen the game scaling the sprite don't scale with other things

+++ src/entities/Player.java
@@ -2,14 +2,18 @@ package entities;
 
 import main.Game;
 import utilz.LoadSave;
-
+import utilz.Constants.PlayerConstants;
+import static utilz.HelpMeMethod.*;
+import javax.imageio.ImageIO;
 import java.awt.*;
+import java.awt.event.KeyEvent;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
+import java.io.InputStream;
 
+
+import static main.Game.GAME_HEIGHT;
 import static utilz.Constants.PlayerConstants.*;
-import static utilz.HelpMeMethod.*;
-
 
 public class Player extends Entity{
     //player attributes
@@ -38,10 +42,11 @@ public class Player extends Entity{
     // store load save = level data
     private int[][] levelData;
 
+
     public Player(float startX, float startY, int width , int height) throws IOException {
         super(startX, startY, width, height);
         loadAnimation();
-        initHitBox(x, y, (int)(28* Game.SCALE),(int)(28* Game.SCALE));        // something wrong here with the hitbox when 32*Game.SCALE
+        initHitBox(x, y, 28* Game.SCALE,28* Game.SCALE);        // something wrong here with the hitbox when 32*Game.SCALE
                                                                             // work fine with 28*Game.SCALE
     }
 
@@ -52,11 +57,13 @@ public class Player extends Entity{
         }
     }
 
+
     private void updateJumpState() {
         if(isInAir){return;}
         isInAir = true;
         airSpeed = BASE_JUMP_FORCE;
     }
+
     public void update() {
         updatePosition();
         updateAnimationTick();
@@ -180,7 +187,6 @@ public class Player extends Entity{
         up = false; // redundant but need for look up
         right = false;
     }
-
     public boolean isLeft() {
         return left;
     }
